Dependencies File/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pyyaml==6.0.2
pandas==2.2.2
streamlit==1.37.0
